refactor-fe: components 폴더명 파일명 수정
컴포넌트 폴더명 파일명 파스칼 케이스로 변경
제일 작은 단위의 순수 ui 컴포넌트들은 components/ui로 이동
