+++ frontend/src/components/menu/contentMenu.tsx
@@ -1,4 +1,4 @@
-import Text from '../text/text';
+import Text from '../ui/Text';
 
 interface ContentMenuProps {
   categorys: string[];
