+++ frontend/src/components/slider/imageSlider.tsx
@@ -1,8 +1,8 @@
 import { useRef, useState } from 'react';
 import LeftArrowIcon from '../../assets/icons/leftArrowIcon.svg?react';
 import RightArrowIcon from '../../assets/icons/rightArrowIcon.svg?react';
-import Text from '../text/text';
-import Image from '../image/image';
+import Text from '../ui/Text';
+import Image from '../ui/Image';
 
 interface ImageSliderProps {
   width?: number;
