+++ frontend/src/components/ui/Image.tsx
